Add first implementation of ADC for NND+BBQ
