+++ cpp/include/cuvs/neighbors/common.hpp
@@ -162,32 +162,6 @@ struct dataset;
 template <typename ContainerType, typename DataT, typename IdxT, typename Accessor>
 struct dataset_view;
 
-/**
- * Storage layout of BBQ/OSQ quantized component codes in each dataset row.
- *
- * `single_bit` Binary-packed bitstream (packAsBinary). Used for bits == 1.
- * `dibit` matches Lucene's transposeDibit. Used for bits == 2.
- * `unsigned_byte` one uint8_t per component. Used for bits == 7 and 8.
- */
-enum class bbq_code_layout {
-  single_bit, /** Each dimension is quantized to a single bit and packed into bytes. Reflects
-               * OptimizedScalarQuantizer.packAsBinary. During query time, the query vector is
-               * quantized to 4 bits per dimension. */
-  dibit,      /** Each dimension is quantized to 2 bits (dibit) and packed into bytes. Reflects
-               * OptimizedScalarQuantizer.transposeDibit. During query time, the query vector is
-               * quantized to 4 bits per dimension. */
-  transpose_half_byte, /** Each dimension is quantized to 4 bits, optimized for bitwise operations.
-                        * Reflects OptimizedScalarQuantizer.transposeHalfByte. the first bit of
-                        * every dimension is in the first set dimensions bits, or (dimensions/8)
-                        * bytes. The second, third, and fourth bits are in the second, third, and
-                        * fourth set of dimensions bits, respectively. Format used for queries. */
-  packed_nibble, /** Each dimension is quantized to 4 bits, two values are packed into each output
-                  * byte. Reflects OffHeapScalarQuantizedVectorValues.packNibbles. */
-  seven_bit,     /** Each dimension is quantized to 7 bits and treated as a signed value. */
-  unsigned_byte, /** Each dimension is quantized to 8 bits and treated as an unsigned value. */
-
-};
-
 namespace detail {
 
 // Default owning/view accessors for public dataset aliases.
@@ -443,129 +417,6 @@ struct vpq_dataset_view_storage {
   }
   [[nodiscard]] owning_dataset_type const& dset() const noexcept { return *dataset_; }
 };
-
-// -----------------------------------------------------------------------------
-// BBQ compressed
-// -----------------------------------------------------------------------------
-
-template <typename CodesMatrixT,
-          typename FloatVectorT,
-          typename SumVectorT,
-          typename CentroidVectorT,
-          typename DataT,
-          typename IdxT>
-struct bbq_dataset_owning_storage {
-  using math_type = DataT;
-
-  CodesMatrixT codes;
-  FloatVectorT lower_intervals;
-  FloatVectorT upper_intervals;
-  FloatVectorT additional_corrections;
-  SumVectorT quantized_component_sums;
-  CentroidVectorT centroid;
-
-  uint32_t bits{};
-  bbq_code_layout layout{bbq_code_layout::single_bit};
-  cuvs::distance::DistanceType metric{cuvs::distance::DistanceType::L2Expanded};
-  float centroid_norm_sq{};
-
-  bbq_dataset_owning_storage(CodesMatrixT&& codes,
-                             FloatVectorT&& lower_intervals,
-                             FloatVectorT&& upper_intervals,
-                             FloatVectorT&& additional_corrections,
-                             SumVectorT&& quantized_component_sums,
-                             CentroidVectorT&& centroid,
-                             uint32_t bits,
-                             bbq_code_layout layout,
-                             cuvs::distance::DistanceType metric,
-                             float centroid_norm_sq) noexcept
-    : codes{std::move(codes)},
-      lower_intervals{std::move(lower_intervals)},
-      upper_intervals{std::move(upper_intervals)},
-      additional_corrections{std::move(additional_corrections)},
-      quantized_component_sums{std::move(quantized_component_sums)},
-      centroid{std::move(centroid)},
-      bits{bits},
-      layout{layout},
-      metric{metric},
-      centroid_norm_sq{centroid_norm_sq}
-  {
-    const auto n_rows = static_cast<int64_t>(codes.extent(0));
-    RAFT_EXPECTS(bits >= 1 && bits <= 8, "BBQ bits must be in [1, 8].");
-    RAFT_EXPECTS(codes.extent(1) == static_cast<int64_t>(encoded_row_length()),
-                 "BBQ code row length does not match dim, bits, and layout.");
-    RAFT_EXPECTS(lower_intervals.extent(0) == n_rows && upper_intervals.extent(0) == n_rows &&
-                   additional_corrections.extent(0) == n_rows &&
-                   quantized_component_sums.extent(0) == n_rows,
-                 "Every BBQ correction array must contain one value per row.");
-  }
-
-  [[nodiscard]] auto n_rows() const noexcept -> IdxT { return codes.extent(0); }
-  [[nodiscard]] auto dim() const noexcept -> uint32_t
-  {
-    return static_cast<uint32_t>(centroid.extent(0));
-  }
-  [[nodiscard]] constexpr auto encoded_row_length() const noexcept -> uint32_t
-  {
-    auto const d = dim();
-    switch (layout) {
-      case bbq_code_layout::single_bit: return (d * bits + 7) / 8;
-      case bbq_code_layout::dibit: return bits * ((d + 7) / 8);
-      case bbq_code_layout::packed_nibble: return (d + 1) / 2;
-      case bbq_code_layout::seven_bit: return d;
-      case bbq_code_layout::unsigned_byte: return d;
-      case bbq_code_layout::transpose_half_byte: return 4 * ((d + 7) / 8);
-    }
-    return 0;
-  }
-};
-
-template <typename ContainerType,
-          typename CodesMatrixViewT,
-          typename FloatVectorViewT,
-          typename SumVectorViewT,
-          typename CentroidVectorViewT,
-          typename DataT,
-          typename IdxT,
-          typename Accessor>
-struct bbq_dataset_view_storage {
-  using owning_storage_type = typename ContainerType::
-    template owning_storage<DataT, IdxT, dataset_owning_accessor_for_view<DataT, Accessor>>;
-  owning_storage_type const* storage_ptr_{nullptr};
-
-  CodesMatrixViewT codes;
-  FloatVectorViewT lower_intervals;
-  FloatVectorViewT upper_intervals;
-  FloatVectorViewT additional_corrections;
-  SumVectorViewT quantized_component_sums;
-  CentroidVectorViewT centroid;
-
-  uint32_t bits{};
-  bbq_code_layout layout{bbq_code_layout::single_bit};
-  cuvs::distance::DistanceType metric{cuvs::distance::DistanceType::L2Expanded};
-  float centroid_norm_sq{};
-
-  bbq_dataset_view_storage(owning_storage_type const* storage_ptr) noexcept
-    : storage_ptr_{storage_ptr},
-      codes{storage_ptr_->codes.view()},
-      lower_intervals{storage_ptr_->lower_intervals.view()},
-      upper_intervals{storage_ptr_->upper_intervals.view()},
-      additional_corrections{storage_ptr_->additional_corrections.view()},
-      quantized_component_sums{storage_ptr_->quantized_component_sums.view()},
-      centroid{storage_ptr_->centroid.view()},
-      bits{storage_ptr_->bits},
-      layout{storage_ptr_->layout},
-      metric{storage_ptr_->metric},
-      centroid_norm_sq{storage_ptr_->centroid_norm_sq}
-  {
-  }
-
-  [[nodiscard]] constexpr auto n_rows() const noexcept -> IdxT { return codes.extent(0); }
-  [[nodiscard]] constexpr auto dim() const noexcept -> uint32_t
-  {
-    return static_cast<uint32_t>(centroid.extent(0));
-  }
-};
 }  // namespace detail
 
 // -----------------------------------------------------------------------------
@@ -626,32 +477,6 @@ struct vpq_dataset_container {
   template <typename MathT, typename IdxT, typename Accessor>
   using view_storage =
     detail::vpq_dataset_view_storage<vpq_dataset_container, MathT, IdxT, Accessor>;
-};
-
-// -----------------------------------------------------------------------------
-// BBQ compressed
-// -----------------------------------------------------------------------------
-
-struct bbq_dataset_container {
-  template <typename DataT, typename IdxT, typename Accessor>
-  using owning_storage =
-    detail::bbq_dataset_owning_storage<detail::dense_owning_matrix<uint8_t, IdxT, Accessor>,
-                                       detail::dense_owning_vector<DataT, IdxT, Accessor>,
-                                       detail::dense_owning_vector<int32_t, IdxT, Accessor>,
-                                       detail::dense_owning_vector<DataT, IdxT, Accessor>,
-                                       DataT,
-                                       IdxT>;
-
-  template <typename DataT, typename IdxT, typename Accessor>
-  using view_storage =
-    detail::bbq_dataset_view_storage<bbq_dataset_container,
-                                     detail::dense_view_matrix<const uint8_t, IdxT, Accessor>,
-                                     detail::dense_view_vector<const DataT, IdxT, Accessor>,
-                                     detail::dense_view_vector<const int32_t, IdxT, Accessor>,
-                                     detail::dense_view_vector<const DataT, IdxT, Accessor>,
-                                     DataT,
-                                     IdxT,
-                                     Accessor>;
 };
 
 template <typename ContainerType, typename DataT, typename IdxT, typename Accessor>
@@ -799,121 +624,6 @@ struct dataset_view<vpq_dataset_container, DataT, IdxT, Accessor>
   using view_storage_type::view_storage_type;
 };
 
-// -----------------------------------------------------------------------------
-// BBQ compressed (view holds non-owning pointer to owning dataset)
-// -----------------------------------------------------------------------------
-
-template <typename DataT, typename IdxT, typename Accessor>
-struct dataset<bbq_dataset_container, DataT, IdxT, Accessor> {
-  using container_type = bbq_dataset_container;
-  using owning_storage_type =
-    typename container_type::template owning_storage<DataT, IdxT, Accessor>;
-  using storage_view_type = typename container_type::
-    template view_storage<DataT, IdxT, detail::dataset_view_accessor_for_owning<DataT, Accessor>>;
-
-  dataset(owning_storage_type&& storage) : storage_vector() { add_storage(std::move(storage)); }
-
-  [[nodiscard]] auto as_dataset_view() const noexcept
-    -> dataset_view<bbq_dataset_container,
-                    DataT,
-                    IdxT,
-                    detail::dataset_view_accessor_for_owning<DataT, Accessor>>
-  {
-    return dataset_view<bbq_dataset_container,
-                        DataT,
-                        IdxT,
-                        detail::dataset_view_accessor_for_owning<DataT, Accessor>>{this};
-  }
-
-  void add_storage(owning_storage_type&& storage)
-  {
-    if (storage_vector.size() != 0) {
-      RAFT_EXPECTS(storage_vector.front().dim() == storage.dim(),
-                   "All storages must have the same dimension");
-      RAFT_EXPECTS(storage_vector.front().n_rows() == storage.n_rows(),
-                   "All storages must have the same number of rows");
-      RAFT_EXPECTS(storage_vector.front().metric == storage.metric,
-                   "All storages must have the same metric");
-      RAFT_EXPECTS(!has_bit_and_layout(storage.bits, storage.layout),
-                   "This bit and layout combination is already in the dataset");
-    }
-    storage_vector.push_back(std::move(storage));
-  }
-  std::vector<owning_storage_type> storage_vector;
-
-  [[nodiscard]] auto n_rows() const noexcept
-  {
-    return storage_vector.size() > 0 ? storage_vector.front().n_rows() : IdxT{0};
-  }
-  [[nodiscard]] auto dim() const noexcept -> uint32_t
-  {
-    return storage_vector.size() > 0 ? storage_vector.front().dim() : uint32_t{0};
-  }
-  bool has_bit_and_layout(uint32_t bits, bbq_code_layout layout) const noexcept
-  {
-    for (const auto& storage : storage_vector) {
-      if (storage.bits == bits && storage.layout == layout) return true;
-    }
-    return false;
-  }
-  [[nodiscard]] auto get_storage_view(uint32_t bits, bbq_code_layout layout) const
-    -> storage_view_type
-  {
-    for (size_t i = 0; i < storage_vector.size(); i++) {
-      if (storage_vector[i].bits == bits && storage_vector[i].layout == layout)
-        return storage_view_type{&storage_vector[i]};
-    }
-    RAFT_FAIL("This bit and layout combination is not in the dataset");
-  }
-  auto storage_size() const noexcept { return storage_vector.size(); }
-  [[nodiscard]] auto get_front_storage_view() const -> storage_view_type
-  {
-    RAFT_EXPECTS(storage_vector.size() != 0, "Storage vector is empty");
-    return storage_view_type{&storage_vector[0]};
-  }
-};
-
-template <typename DataT, typename IdxT, typename Accessor>
-struct dataset_view<bbq_dataset_container, DataT, IdxT, Accessor> {
-  using container_type    = bbq_dataset_container;
-  using dataset_type      = dataset<bbq_dataset_container,
-                                    DataT,
-                                    IdxT,
-                                    detail::dataset_owning_accessor_for_view<DataT, Accessor>>;
-  using view_storage_type = typename dataset_type::storage_view_type;
-
-  dataset_type const* dataset_ptr_{nullptr};
-  dataset_view(dataset_type const* dataset_ptr) : dataset_ptr_{dataset_ptr} {}
-  [[nodiscard]] auto n_rows() const noexcept
-  {
-    return dataset_ptr_ != nullptr ? dataset_ptr_->n_rows() : IdxT{0};
-  }
-  [[nodiscard]] auto dim() const noexcept -> uint32_t
-  {
-    return dataset_ptr_ != nullptr ? dataset_ptr_->dim() : uint32_t{0};
-  }
-  [[nodiscard]] auto dset() const noexcept -> dataset_type const& { return *dataset_ptr_; }
-  bool has_bit_and_layout(uint32_t bits, bbq_code_layout layout) const noexcept
-  {
-    return dataset_ptr_ != nullptr ? dataset_ptr_->has_bit_and_layout(bits, layout) : false;
-  }
-  [[nodiscard]] auto get_storage_view(uint32_t bits, bbq_code_layout layout) const
-    -> view_storage_type
-  {
-    RAFT_EXPECTS(dataset_ptr_ != nullptr, "Dataset pointer is nullptr");
-    return dataset_ptr_->get_storage_view(bits, layout);
-  }
-  [[nodiscard]] auto get_front_storage_view() const -> view_storage_type
-  {
-    RAFT_EXPECTS(dataset_ptr_ != nullptr, "Dataset pointer is nullptr");
-    return dataset_ptr_->get_front_storage_view();
-  }
-  auto storage_size() const noexcept
-  {
-    return dataset_ptr_ != nullptr ? dataset_ptr_->storage_size() : std::size_t{0};
-  }
-};
-
 /**
  * @brief Aliases for concrete `dataset` / `dataset_view` layouts.
  */
@@ -981,27 +691,6 @@ template <typename DataT, typename IdxT>
 using host_vpq_dataset_view =
   dataset_view<vpq_dataset_container, DataT, IdxT, detail::host_view_accessor<DataT>>;
 
-template <typename IdxT>
-using device_bbq_dataset =
-  dataset<bbq_dataset_container, float, IdxT, detail::device_owning_accessor<float>>;
-
-template <typename IdxT>
-using device_bbq_dataset_view =
-  dataset_view<bbq_dataset_container, float, IdxT, detail::device_view_accessor<float>>;
-
-template <typename IdxT>
-using device_bbq_dataset_storage_view =
-  dataset_view<bbq_dataset_container, float, IdxT, detail::device_view_accessor<float>>::
-    view_storage_type;
-
-template <typename IdxT>
-using host_bbq_dataset =
-  dataset<bbq_dataset_container, float, IdxT, detail::host_owning_accessor<float>>;
-
-template <typename IdxT>
-using host_bbq_dataset_view =
-  dataset_view<bbq_dataset_container, float, IdxT, detail::host_view_accessor<float>>;
-
 // Maps a dataset view type to its owning (allocating) dataset counterpart.
 // Used by serialize/deserialize to type the out_dataset output parameter;
 // adding a new dataset type only requires adding a new specialization here.
@@ -1031,11 +720,6 @@ struct owning_dataset_for_view<host_standard_dataset_view<DataT, IdxT>> {
 template <typename DataT, typename IdxT>
 struct owning_dataset_for_view<device_vpq_dataset_view<DataT, IdxT>> {
   using type = device_vpq_dataset<DataT, IdxT>;
-};
-
-template <typename IdxT>
-struct owning_dataset_for_view<device_bbq_dataset_view<IdxT>> {
-  using type = device_bbq_dataset<IdxT>;
 };
 
 template <typename DatasetViewT>
@@ -1077,19 +761,6 @@ struct is_vpq_dataset<dataset<vpq_dataset_container, DataT, IdxT, Accessor>> : s
 
 template <typename DatasetT>
 inline constexpr bool is_vpq_dataset_v = is_vpq_dataset<DatasetT>::value;
-
-template <typename DatasetT>
-struct is_bbq_dataset : std::false_type {};
-
-template <typename DataT, typename IdxT, typename Accessor>
-struct is_bbq_dataset<dataset<bbq_dataset_container, DataT, IdxT, Accessor>> : std::true_type {};
-
-template <typename DataT, typename IdxT, typename Accessor>
-struct is_bbq_dataset<dataset_view<bbq_dataset_container, DataT, IdxT, Accessor>> : std::true_type {
-};
-
-template <typename DatasetT>
-inline constexpr bool is_bbq_dataset_v = is_bbq_dataset<DatasetT>::value;
 
 // -----------------------------------------------------------------------------
 // Dataset view compile-time classification (replaces runtime std::variant dispatch).
@@ -1144,11 +815,6 @@ struct dataset_view_kind_of<dataset_view<vpq_dataset_container, MathT, IdxT, Acc
                 "VPQ dataset_view_kind_of expects MathT to be half or float");
   static constexpr dataset_view_kind value =
     std::is_same_v<MathT, half> ? dataset_view_kind::vpq_f16 : dataset_view_kind::vpq_f32;
-};
-
-template <typename DataT, typename IdxT, typename Accessor>
-struct dataset_view_kind_of<dataset_view<bbq_dataset_container, DataT, IdxT, Accessor>> {
-  static constexpr dataset_view_kind value = dataset_view_kind::bbq;
 };
 
 template <typename V>
@@ -1244,18 +910,6 @@ inline constexpr bool is_host_vpq_dataset_view_v =
 template <typename V>
 inline constexpr bool is_vpq_dataset_view_v =
   is_device_vpq_dataset_view_v<V> || is_host_vpq_dataset_view_v<V>;
-
-template <typename V>
-inline constexpr bool is_device_bbq_dataset_view_v =
-  dataset_view_kind_v<V> == dataset_view_kind::bbq && dataset_view_is_device_accessible_v<V>;
-
-template <typename V>
-inline constexpr bool is_host_bbq_dataset_view_v =
-  dataset_view_kind_v<V> == dataset_view_kind::bbq && !dataset_view_is_device_accessible_v<V>;
-
-template <typename V>
-inline constexpr bool is_bbq_dataset_view_v =
-  is_device_bbq_dataset_view_v<V> || is_host_bbq_dataset_view_v<V>;
 
 /** True for any device-resident dataset view. */
 template <typename V>
